Add GC.@ preserve where pointer() is used
Using the value returned by `pointer()` is unsafe unless Julia is told
to preserve the owner of that memory. GC.@ preserve is the right way to
do this.

Also set the pointer fields `next_in` and `next_out` to `C_NULL`
whenever we're finished with them to prevent use-after-free surprises.
While not exactly necessary in an ideal bug-free world, this should help
with robustness by detecting any accidental reuse of stale buffers.