Back to CodecZlib
Turns out it truly is the lesser of two evils. While great care has to
be taken to make sure the TranscodingStreams are in a good state
throughout all the reading and writing, it's easier to do that and use
the dual buffers than it is to try to determine how much data has been
written to a compressed stream or how much uncompressed data has been
read from a compressed stream.