Fix validation of files with data descriptors
Data descriptors are difficult to validate, but we can assume that if
the SentinelLimiter stopped reading, it ended with a proper data
descriptor.

This needs to be fixed because the code currently trusts the compression
codec to signal end of stream when it sees an EOF marker in the data, so
SentinelLimiter's checks are bypassed.