Fix SentinelLimiter tests
SentinelLimiter is careful to keep track of the expected type used to
describe content lengths. The test was enforcing UInt64 lengths in the
SentinelLimiter, but when run on a 32-bit system, sizeof(CONTENT) would
return an Int32. Because the test was manually creating a Sentinel block
by writing to an IO object with write(io, value), a 32-bit number was
being written, but the SentinelLimiter was reading off a 64-bit number
and reaching EOF too soon.